mon sprint java

+++ Pidev/src/Argents/Entitys/Argent.java
@@ -14,19 +14,19 @@ import java.sql.Date;
 public class Argent {
     
     private int id;
-    private float montant;
+    private int montant;
     private Date date;
 
     public Argent() {
     }
 
-    public Argent(int id, float montant, Date date) {
+    public Argent(int id, int montant, Date date) {
         this.id = id;
         this.montant = montant;
         this.date = date;
     }
     
-    public Argent(float montant, Date date) {
+    public Argent(int montant, Date date) {
         this.montant = montant;
         this.date = date;
     }
@@ -47,7 +47,7 @@ public class Argent {
         this.id = id;
     }
 
-    public void setMontant(float montant) {
+    public void setMontant(int montant) {
         this.montant = montant;
     }
